🚧新增选课中心

+++ server/java-server/src/main/java/org/fatmansoft/teach/controllers/BaseController.java
@@ -23,7 +23,6 @@ import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBo
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.awt.*;
 import java.io.*;
 import java.lang.reflect.Method;
 import java.nio.charset.StandardCharsets;
@@ -162,12 +161,13 @@ public class BaseController {
     @PreAuthorize("hasRole('ADMIN')")
     public DataResponse menuDelete(@Valid @RequestBody DataRequest dataRequest) {
         Integer id = dataRequest.getInteger("id");
-        List<MenuInfo> children = menuInfoRepository.findMenuInfosByPid(id);
-        Optional<MenuInfo> op = menuInfoRepository.findById(id);
-        if (op.isPresent()){
-            menuInfoRepository.delete(op.get());
-            menuInfoRepository.deleteAll(children);
+        int count  = menuInfoRepository.countMenuInfoByPid(id);
+        if(count > 0) {
+            return CommonMethod.getReturnMessageError("存在子菜单，不能删除！");
         }
+        Optional<MenuInfo> op = menuInfoRepository.findById(id);
+        if (op.isPresent())
+            menuInfoRepository.delete(op.get());
         return CommonMethod.getReturnMessageOK();
     }
